Code example under point 5 would not compile. (#41)
* Code example under point 5 would not compile.

Thanks for your great work! I found a few issues in the README code example for which I'd like to propose a solution.

The code example provided under point 5 had a few problems.

Function call and arguments:
- missing `,` after `onDrag` closure
- `header` closure has a default value and can thus be omitted, but cannot be empty.

Custom extension used:
`.cornerRadius(12, corners: [.topLeft, .topRight])` is not part of the Swift standard library. So unless this extension is provided as part of BottmSheet, which I wouldn't recommend, this will cause an error.

As of iOS 16 we would have the option to use:
```Swift
.clipShape(
  .rect(
    topLeadingRadius: 12,
    bottomLeadingRadius: 0,
    bottomTrailingRadius: 0,
    topTrailingRadius: 12
  )
)
```

I didn't add it to my pr because of the limited backwards compatibility.

* Fix indentation